Add "secondary" and "ambivalent" spam folders
For automated systems that move spam around (like a spamassassin + sieve
setup), there may be several folders that are for spam. For example, a
folder for spam classified by spamassassin with high confidence, one for
low confidence, and one for mails marked / moved by the user's mail
client (like the adaptive spam filter in thunderbird).

This commit adds config options to configure these folders and updates
the hide/unhide logic for the mark as spam / mark as not spam buttons
accordingly: The secondary spam folders show the "Not Junk" button,
while "suspicious" folders show both buttons.